requirements.txt
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
decorator==5.1.1
Flask==2.1.2
Flask-JWT-Extended==4.4.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.0
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
validators==0.19.0
Werkzeug==2.1.2
zipp==3.8.0
